Update README.md
# Image-Encryptor-and-Decryptor
An image encrypting and decrypting software. 
Language used-Java
IDE used- Visual Studio code


The image encryptor and decryptor can be used to encrypt any image using a key. 
Select the image which one wants to encrypt by providing a key (number) which will be further used to decrypt the file, Select the file by clicking on file location and selecing the file which one wants to encrypt.
The image will be encrypted.
For decryption of the image. Input the key and open the file location and select the image which has been encrypted. The image will be decrypted.